Add support for worker initialization functions
Some workers may require an initialization sequence before they should
connect to the socket.

In order to utilize this, simply add:
  .set('worker init', 'method name')
to your cluster instance.

'method name' must be a method of the server that takes one callback
argument as follows:

function server.worker_init(callback) {
  <initialization code goes here>
  if (error) {
    return callback(<error>)
  }
  // no error
  return callback(null);
}

When running the callback, one argument should be supplied. If the
<error> is "truthy", then the worker will consider it to have failed. If
it is "falsy", then the worker will connect to the master.

Signed-off-by: Joshua Rubin <joshua@rubixconsulting.com>